Numerous additions poorly organized. >_<
Removed the "minimised" file and added a Makefile to build it. Requires
Google's Closure Compiler, which I've called 'jsc' for shortness. That
name might conflict with some program unknown to me so the Makefile
probably could use some improvements to be more cross-platform. For now,
I think I'm the only ojne using it anyway so no harm done. ^_^ The
minified file is optional anyway... And you're free to use whatever you
want instead.

Modified the navigation menu to be wider to accept so many more options
without jolting the page content. Originally, I had added a second row
of options to the navigation menu, but I found on first load for a page
with new content the page loading combined with the JavaScript's
execution caused a noticable delay in the drawing, which would often
result in me attempting to click a link right when the page jolts down,
often causing me to click the wrong link. Very annoying.

Added a link to the outbox of the navigation menu.

Also added support for the <br /> tag, which I guess I forgot to add to
the clonePost method. I'm still getting the annoying xmlns attribute in
the output so I tried adding code to remove it also, but it still
doesn't seem to be working. Haven't bothered to figure out where in the
pipeline that attribute is added. Also added code to trim the text
nodes, though I'm not yet sure whether it works as intended or has
negative effects. It's possible it will remove unintended whitespace or
might not work at all. Needs testing, but I'm satisfied committing it
for now. >_>

Working with Subversion at j0rb has me used to not branching...
Something that I'd like Git to help me overcome. Apologies for the messy
commit. >_>