Modifying forum links; default to last read.
Now that I'm using Vimperator, I tend to navigate with the keyboard alone.
I'm currently unaware of any way to 'double-click' something using
Vimperator, so I can't activate Allegro.cc's "jump to last read post"
mechanism without grabbing the mouse and double-clicking the background of
a thread in the list. To save myself the trouble of first navigating to
the page and then navigating to the last read, something that proves
inconsistent anyway, I decided to use acc.js to modify the links to
default to the last read post. This will be convenient for Vimperator uses
and maybe even normal users. As a fall back, I've also added a normal link
to the right of the column for each thread with a default link labelled
"Top". This will allow you to to still open the thread as per normal, to
the top of the page, if you have some reason to (i.e., re-reading the
thread or what have you).

Since there is yet to be a user-specific configuration for acc.js (since
I'm the only known user anyway), the only way to currently disable this is
to comment out the block of code (or alternatively surround it with a
failing condition, like if(false) { ... }. Let me know if you'd like
something more convenient (or if anyone else is even using this!).