Modify ignore feature to not change page layout.
Changing the height of things dynamically is screwing up the browser's
ability to jump to the last read part automagically, which I heavily rely
on.

Instead, I attempt to just make the ignored posts invisible/hidden, and
then try to layer the 'show post' "bar" on top of it... I don't think I
have the positioning right, but I can't really explain what's happening.
:'( It seems to work at first glance, but I'll have to browse A.cc a bit
to see that the browser is positioning me properly again before I'm
convinced that this bug is "fixed"...