Improve wording about maintaining notice.
The only intent is really to help people to find the original script and
know their rights with it, and to know what changes have been made by
third parties. Don't feel restricted in any way. You are pretty much
free to do as you please as long as you are honest.